// backend/src/types.ts
import { z } from "zod";

export const SignInSchema = z.object({
    email: z.string().email(),
    password: z.string().min(6), 
});

export const SignUpSchema = z.object({
    name: z.string().min(1), 
    email: z.string().email(),
    password: z.string().min(6),
});

export type SignInType = z.infer<typeof SignInSchema>;
export type SignUpType = z.infer<typeof SignUpSchema>;